Rename activity2 to activity
