+++ packages/component/src/Activity/Avatar.js
