+++ packages/component/src/Activity/Bubble.js
