+++ packages/component/src/Activity/CarouselFilmStrip.js
