+++ packages/component/src/Activity/CarouselLayout.js
