+++ packages/component/src/Activity/SendStatus.js
