+++ packages/component/src/Activity/Speak.js
