+++ packages/component/src/Activity/StackedLayout.js
