+++ packages/component/src/Activity/Timestamp.js
