+++ packages/component/src/BasicTranscript.js
@@ -5,7 +5,7 @@ import React from 'react';
 import ScrollToBottom from 'react-scroll-to-bottom';
 
 import connectWithContext from './connectWithContext';
-import SpeakActivity from './Activity2/Speak';
+import SpeakActivity from './Activity/Speak';
 
 const ROOT_CSS = css({
   display: 'flex',
